% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw1_get_csv.R
\name{get_hw1_airline_df}
\alias{get_hw1_airline_df}
\title{Load set of HW1 dataframes}
\usage{
get_hw1_airline_df(login)
}
\arguments{
\item{login}{your vle login}
}
\value{
csv file that will be saved loaded to your disk and environment
}
\description{
This function loads a file in your environment
}
