% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw1_get_csv.R
\name{get_hw1_airport_df}
\alias{get_hw1_airport_df}
\title{Load set of HW1 dataframe with airports}
\usage{
get_hw1_airport_df(login)
}
\arguments{
\item{login}{your vle login}
}
\value{
dataframe loaded to your environment
}
\description{
This function loads a file in your environment
}
