% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw1.R
\name{get_hw1_csv}
\alias{get_hw1_csv}
\title{Load a HW1}
\usage{
get_hw1_csv(login = Sys.info()["user"])
}
\arguments{
\item{login}{your vle login}
}
\value{
csv file that will be saved loaded to your disk and environment
}
\description{
This function loads a file as a csv.
}
