% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw1.R
\name{get_hw1_data}
\alias{get_hw1_data}
\title{Load set of HW1 dataframe with airlines}
\usage{
get_hw1_data(login, path = NULL)
}
\arguments{
\item{login}{your vle login}

\item{path}{place to write csv could be path, i.e. ~/my_folder/, or null}
}
\value{
list of dataframes, also write csv files to your disk
}
\description{
This function loads a file in your environment
}
