% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw1_get_questions.R
\name{get_hw1_questions}
\alias{get_hw1_questions}
\title{Get a HW1 questions}
\usage{
get_hw1_questions(login = Sys.info()["user"])
}
\arguments{
\item{login}{your vle login}
}
\value{
df that you read from a directory
}
\description{
This function loads questions in df
}
