% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw_ind.R
\name{get_hw_ind_questions}
\alias{get_hw_ind_questions}
\title{Get a HW1 questions}
\usage{
get_hw_ind_questions(login)
}
\arguments{
\item{login}{your vle login}
}
\value{
df that you read from a directory
}
\description{
This function loads questions in df
}
