% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab06_emb.R
\name{get_lab06_glove_embeddings}
\alias{get_lab06_glove_embeddings}
\title{Load set of lab06 functions to save r2 server from students}
\usage{
get_lab06_glove_embeddings(login, word = "paris")
}
\arguments{
\item{word}{that you want to vectorize}
}
\value{
matrix with embeddings from glove
}
\description{
This function loads a file in your environment
}
