% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab06_emb.R
\name{get_lab06_glove_sim}
\alias{get_lab06_glove_sim}
\title{Load set of HW1 dataframe with airports}
\usage{
get_lab06_glove_sim(word = "paris", n = 5)
}
\arguments{
\item{word}{that you want to vectorize}

\item{n}{number of words to return}
}
\value{
dataframe loaded to your environment
}
\description{
This function loads a file in your environment
}
