% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw1_read_csv.R
\name{open_hw1_csv}
\alias{open_hw1_csv}
\title{Open a HW1 csv}
\usage{
open_hw1_csv(login = Sys.info()["user"])
}
\arguments{
\item{login}{your vle login}
}
\value{
df that you read from a directory
}
\description{
This function loads a file as a csv.
}
