% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw1.R
\name{get_hw1_data}
\alias{get_hw1_data}
\title{Load set of HW1 dataframe with airlines}
\usage{
get_hw1_data(login, df_name = "airline")
}
\arguments{
\item{login}{your vle login}

\item{df_name}{name of a dataset airline, airport, lounge or seat}
}
\value{
list of dataframes, also write csv files to your disk
}
\description{
This function loads a file in your environment
}
