% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw_ind.R
\name{get_hw_ind_questions}
\alias{get_hw_ind_questions}
\title{Get a HW1 questions}
\usage{
get_hw_ind_questions(login)
}
\arguments{
\item{login}{your vle login}
}
\value{
text of the question
}
\description{
This function loads questions for the ind task
}
